+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
+- added: (XRP) Add mnemonic import support
 - added: (XRP) Add destination tag to processed transactions
+- changed: (XRP/XLM) Create mnemonics by default
 - changed: (XRP) Check recipient trust lines before sending token
 - fixed: `parseUri` case sensitivity when looking for matching builtin token contract addresses
 - removed: Deprecate Binance Beacon Chain